test: unifying tests
